requirements.txt
flask
pymongo
aiohttp
beautifulsoup4
pyOpenSSL
mmh3
requests
dnspython
